# openapi.yaml
components:
  schemas:
...
    Body_login_api_v1_auth_login_post:
      properties:
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Secret
        grant_type:
          anyOf:
          - pattern: password
            type: string
          - type: 'null'
          title: Grant Type
        password:
          title: Password
          type: string
        scope:
          default: ''
          title: Scope
          type: string
        username:
          title: Username
          type: string
      required:
      - username
      - password
      title: Body_login_api_v1_auth_login_post
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Token:
      properties:
        access_token:
          title: Access Token
          type: string
        token_type:
          title: Token Type
          type: string
      required:
      - access_token
      - token_type
      title: Token
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    OAuth2PasswordBearer:
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
      type: oauth2
info:
  description: Backend API for cybercrime investigation dashboard
  title: Cybercrime Investigation API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/auth/login:
    post:
      description: Login and get access token
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_api_v1_auth_login_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Login
      tags:
      - authentication
...
  /api/v1/dashboard/activity-feed:
    get:
      description: 'Combined activity stream: evidence uploads, complaints, wallet
        freeze/unfreeze, AI reports, etc.

        Currently derived from AuditLog; can be extended later.'
      operationId: get_activity_feed_api_v1_dashboard_activity_feed_get
      parameters:
      - in: query
        name: actor_id
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Actor Id
      - in: query
        name: type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      - in: query
        name: start
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Start
      - in: query
        name: end
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: End
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Activity Feed
      tags:
      - dashboard
  /api/v1/dashboard/notifications:
    get:
      description: 'Aggregated notifications: AI, complaints, wallets, system.

        Filtered by user role:

        - Superadmin: sees all notifications including access requests

        - Investigator: sees only their own activity and messages'
      operationId: get_notifications_api_v1_dashboard_notifications_get
      parameters:
      - in: query
        name: type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
      - in: query
        name: severity
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - OAuth2PasswordBearer: []
      summary: Get Notifications
      tags:
      - dashboard
  /api/v1/dashboard/notifications/mark-all-read:
    post:
      description: 'Mark all notifications as read for the current user.

        Updates User.last_notification_read_at and marks all Messages as read.'
      operationId: mark_notifications_read_api_v1_dashboard_notifications_mark_all_read_post
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
      summary: Mark Notifications Read
      tags:
      - dashboard
  /api/v1/dashboard/priority-queue:
    get:
      description: 'Build a unified queue of complaints + high-risk incident reports,

        call OpenRouter for AI priority scoring, and return sorted list.'
      operationId: get_priority_queue_api_v1_dashboard_priority_queue_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Priority Queue
      tags:
      - dashboard
  /api/v1/dashboard/risk-trends:
    get:
      description: 'Compute risk trends from incident reports (risk levels) and complaints
        (regional volume)

        over the last N days.'
      operationId: get_risk_trends_api_v1_dashboard_risk_trends_get
      parameters:
      - in: query
        name: days
        required: false
        schema:
          default: 30
          maximum: 365
          minimum: 1
          title: Days
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Risk Trends
      tags:
      - dashboard
...